Break hero out into component
